move local query info
